Add domainMax option to column charts
If set, only columns whose value equals domainMax are rendered with the
full columnHeight. Columns whose value exceeds domainMax are rendered
outside of the chart.

Support domainMax in stacked column charts.

Pass the domainMax value when creating axes. The createAxes mixin does
not read it from the instance object. The value is automatically used
as the default end value for range based axes.