Enable strict null checks (#16)
* Enable strict null checks

* Deal with Map getters possibly returning undefined

TypeScript can't infer that we've already set a the initial value if the
item was missing from the map. Instead fall back to a new initial value
when getting the item, and add it to the map after checking whether it
had an item at all.

* Silence strict null checks for optional instance method access

The method must be available if the selector in question is not provided
as an option. This is not something that can be verified at design time,
so silence the error and let the runtime throw an exception instead.

* Only return optional property if it's available

TypeScript doesn't let us return a possibly undefined value for optional
properties, we can only return the expected value.

* Declare optional arguments last

This seems more sensible then declaring them with possible undefined values.

* Explicitly declare variables as possibly being undefined

* Explicit undefined comparison rather than in-check

TypeScript CFA doesn't work with in-checks.

* Silence faux-Invalidatable strict null check violation

The way these mixins are created isn't ideal. It should be better once
<https://github.com/dojo/compose/issues/74> is in place. Silence the
violation for now.

* Don't needlessly assign null value

* Assign optional option fallbacks to new variables

This way TypeScript correctly believes the variable is never undefined.

* Type possible null value as such